console: Move application logics to new file move logics to logics/applications.js
